Move events enums to the package's init file
