+++ saleor/graphql/order/enums.py
@@ -1,6 +1,6 @@
 import graphene
 
-from ...events.types import OrderEvents, OrderEventsEmails
+from ...events import OrderEvents, OrderEventsEmails
 
 OrderEventsEnum = graphene.Enum.from_enum(OrderEvents)
 OrderEventsEmailsEnum = graphene.Enum.from_enum(OrderEventsEmails)
