+++ saleor/graphql/order/resolvers.py
@@ -1,8 +1,8 @@
 import graphene
 import graphene_django_optimizer as gql_optimizer
 
+from ...events import OrderEvents
 from ...events.models import OrderEvent
-from ...events.types import OrderEvents
 from ...order import OrderStatus, models
 from ...order.utils import sum_order_totals
 from ..utils import filter_by_period, filter_by_query_param
