+++ saleor/payment/utils.py
@@ -13,7 +13,7 @@ from ..account.models import Address, User
 from ..checkout.models import Checkout
 from ..core import analytics
 from ..events.models import OrderEvent
-from ..events.types import OrderEventsEmails
+from ..events import OrderEventsEmails
 from ..order import utils as order_utils
 from ..order.emails import send_payment_confirmation
 from ..order.models import Order
