+++ tests/api/test_homepage.py
@@ -1,4 +1,4 @@
-from saleor.events.types import OrderEvents
+from saleor.events import OrderEvents
 
 from .utils import get_graphql_content
