+++ tests/conftest.py
@@ -26,7 +26,7 @@ from saleor.discount import VoucherType
 from saleor.discount.models import Sale, Voucher, VoucherTranslation
 from saleor.menu.models import Menu, MenuItem
 from saleor.events.models import OrderEvent
-from saleor.events.types import OrderEvents
+from saleor.events import OrderEvents
 from saleor.order import OrderStatus
 from saleor.order.models import FulfillmentStatus, Order
 from saleor.order.utils import fulfill_order_line, recalculate_order
