+++ tests/test_payment.py
@@ -6,7 +6,7 @@ from django.conf import settings as dj_settings
 from django.core.exceptions import ImproperlyConfigured
 from django.template.loader import get_template
 
-from saleor.events.types import OrderEvents, OrderEventsEmails
+from saleor.events import OrderEvents, OrderEventsEmails
 from saleor.order.views import PAYMENT_TEMPLATE
 from saleor.payment import (
     ChargeStatus, GatewayError, OperationType, PaymentError, TransactionKind,
